Add config support for randomized attachments
This addition to the RVConfig should work in tandem with the new engine changes to support attachment randomization when weapons are spawned on the map. Currently this is limited in scope to magazines and optics, but can be expanded for handguards, flashlights, and buttstocks. AK74 variants should now all have a chance to spawn a random attachment, will test this soon.